% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_view.R
\name{download_view}
\alias{download_view}
\title{Publish a R data frame as a Tableau published data source}
\usage{
download_view(
  server,
  name,
  file,
  workbook_name = NULL,
  filters = list(),
  parameters = list(),
  options = list()
)
}
\arguments{
\item{server}{A Tableau server object. See `tableau_connect()`.}

\item{name}{The name of the view to download.}

\item{file}{The file path to download to. Must have an extension of ".csv", ".png", or ".pdf".}

\item{workbook_name}{The name of the workbook that the view is within. Use if other views in the server share the same name. Optional.}

\item{filters}{A named list of items each having length 1 where the names represent the names of the fields to be filtered on and the values represent the items to keep. Optional.}

\item{parameters}{A named list of items each having length 1 where the names represent the names of the parameters and the values represent the parameter values. Optional.}

\item{options}{A named list of options to be passed to methods. See `download_options` for a data frame of all options available.}
}
\description{
Requires reticulate and Python interpreter
}
\examples{
/dontrun{

keyring::key_set("tableau", "nageld1")

server <- tableau_connect("mdh_internal_dev",
username = "nageld1",
password = keyring::key_get("tableau", "nageld1")
)

view_name <- "Health Indicators"
workbook_name <- "World Indicators mntableau"
file <- tempfile(fileext = ".pdf")
filters <- list(c("Region" = "The Americas"))

download_view(
  server,
  name = view_name,
  file = file,
  workbook_name = workbook_name,
  filters = filters
)
}

}
