% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_workbook.R
\name{download_workbook}
\alias{download_workbook}
\title{Download a workbook from a Tableau site}
\usage{
download_workbook(conn, name, file, project_name = NULL, format = NULL)
}
\arguments{
\item{conn}{A Tableau connection object. See `tableau_connect()`.}

\item{name}{The name of the workbook. Case sensitive.}

\item{file}{The name of the file to output to.}

\item{project_name}{The name of the project that the workbook is in. Use if the workbook name is not unique on your Tableau site. Optional.}

\item{format}{One of "tbwx", "twb", "pdf", or "ppt". Defaults to"twbx" if NULL and no extension is provided in `file`.}
}
\description{
Requires reticulate and Python interpreter
}
\examples{
/dontrun{

keyring::key_set("tableau", "nageld1")

conn <- tableau_connect("mdh_internal_dev",
username = "nageld1",
password = keyring::key_get("tableau", "nageld1")
)

get_project_id(conn, "MNIT-MDH")
get_project_id(conn, "Newborn", parent_name = "PHL")
}

}
