% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_id.R
\name{get_project_id}
\alias{get_project_id}
\title{Get the project id for a given project name and optional parent project}
\usage{
get_project_id(server, name, parent_name = NULL)
}
\arguments{
\item{server}{A Tableau server object. See `tableau_connect()`.}

\item{name}{The name of the project. Case sensitive.}

\item{parent_name}{The name of the parent project of `name`. Use if `name` is not a unique project name on your Tableau site. Optional.}
}
\description{
Requires reticulate and Python interpreter
}
\examples{
/dontrun{

keyring::key_set("tableau", "nageld1")

server <- tableau_connect("mdh_internal_dev",
username = "nageld1",
password = keyring::key_get("tableau", "nageld1")
)

get_project_id(server, "MNIT-MDH")
get_project_id(server, "Newborn", parent_name = "PHL")
}

}
