% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_data_frame.R
\name{publish_data_frame}
\alias{publish_data_frame}
\title{Publish a R data frame as a Tableau published data source}
\usage{
publish_data_frame(
  data,
  server,
  name,
  project_name,
  parent_project_name = NULL,
  datasource_description = NULL
)
}
\arguments{
\item{data}{A data frame that can be published as a data source.}

\item{server}{A Tableau server object. See `tableau_connect()`.}

\item{name}{The name of the published data source.}

\item{project_name}{The name of the project to publish to (must exist already). Case sensitive.}

\item{parent_project_name}{The name of the parent project of `project_name`. Use if `project_name` is not a unique project name on your Tableau site. Optional.}

\item{datasource_description}{A string describing the published data source. Optional.}
}
\description{
Requires arrow package
}
\examples{
/dontrun{

keyring::key_set("tableau", "nageld1")

server <- tableau_connect("mdh_internal_dev",
username = "nageld1",
password = keyring::key_get("tableau", "nageld1")
)

publish_data_frame(tableau_sites, server,
name = "MN Tableau sites",
project_name = "mntableau data sources",
parent_project_name = "Tableau REST API testing",
datasource_description = "MN state government Tableau site info"
)
}

}
