% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau_connect.R
\name{tableau_connect}
\alias{tableau_connect}
\title{Open a server connection to a MN state Tableau site}
\usage{
tableau_connect(
  site_name,
  username = NULL,
  password = NULL,
  token_name = NULL,
  token_secret = NULL
)
}
\arguments{
\item{site_name}{Name of the Tableau site you want to connect to. Use `tableau_sites` to view valid sites.}

\item{username}{Your Tableau server username. Do not include if using a token.}

\item{password}{Your Tableau server password. Recommend using the `keyring` package to set your password. Do not include if using a token.}

\item{token_name}{Your Tableau server token name. Do not include if using a username/password.}

\item{token_secret}{Your Tableau server token secret. Recommend using the `keyring` package to set your token secret. Do not include if using a username/password.}
}
\description{
Requires reticulate and Python interpreter
}
\examples{
/dontrun{

keyring::key_set("tableau", "nageld1")

server <- tableau_connect("mdh_internal_dev",
username = "nageld1",
password = keyring::key_get("tableau", "nageld1")
)
}

}
