% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_view.R
\name{download_view}
\alias{download_view}
\title{Publish a R data frame as a Tableau published data source}
\usage{
download_view(
  server,
  name,
  file,
  workbook_name = NULL,
  filters = list(),
  parameters = list(),
  options = list()
)
}
\arguments{
\item{server}{A Tableau server object. See `tableau_connect()`.}

\item{name}{The name of the view to download.}

\item{file}{The file path to download to. Must have an extension of ".csv", ".png", or ".pdf".}

\item{workbook_name}{The name of the workbook that the view is within. Use if other views in the server share the same name. Optional.}

\item{filters}{A list of named items each having length 1 where the names represent the names of the fields to be filtered on and the values represent the items to keep. Optional.}

\item{parameters}{A list of named items each having length 1 where the names represent the names of the parameters and the values represent the parameter values. Optional.}
}
\description{
Requires reticulate and Python interpreter
}
\examples{
/dontrun{

keyring::key_set("tableau", "nageld1")

server <- tableau_connect("mdh_internal_dev",
username = "nageld1",
password = keyring::key_get("tableau", "nageld1")
)

file <- "C:/Users/nageld1/Downloads/test.pdf"

filters <- list(c("Customer Name" = "Aaron Hawkins"))

download_view(server, "Subtotals example", file, filters = filters)
}

}
